- Deprecated the Hung Routine Handler action module. - Refined the entire routine system to create each routine instance when a task starts. - Changed the default routine status of tasks to be `Ready` from `Inactive`. - Fixed a bug that the wizard form field of the Exit Code occurrence type did not function as of the previous version. - Fixed a bug that when updating module options, the last run time meta value got lost.

+++ include/class/module/occurrence/volatile/TaskScheduler_Occurrence_Volatile.php
@@ -4,9 +4,9 @@
  * 
  * @package     Task Scheduler
  * @copyright   Copyright (c) 2014, Michael Uno
- * @author		Michael Uno
- * @authorurl	http://michaeluno.jp
- * @since		1.0.0
+ * @author        Michael Uno
+ * @authorurl    http://michaeluno.jp
+ * @since        1.0.0
  */
 
 /**
@@ -16,36 +16,42 @@
  * 
  */
 class TaskScheduler_Occurrence_Volatile extends TaskScheduler_Occurrence_Base {
-		
-	/**
-	 * The user constructor.
-	 */
-	public function construct() {}
-	
-	/**
-	 * Returns the label for the slug.
-	 */
-	public function getLabel( $sSlug ) {
-		return __( 'Volatile', 'task-scheduler' );
-	}			
+        
+    /**
+     * The user constructor.
+     */
+    public function construct() {}
+    
+    /**
+     * Returns the label for the slug.
+     */
+    public function getLabel( $sSlug ) {
+        return __( 'Volatile', 'task-scheduler' );
+    }            
 
-	/**
-	 * Returns the description of the module.
-	 */
-	public function getDescription( $sDescription ) {
-		return __( 'Triggers actions only once, used mostly by threads.', 'task-scheduler' );
-	}		
-	
-	/**
-	 * Deletes stored tasks.
-	 */
-	public function doAfterAction( $oRoutine, $sExitCode ) {
+    /**
+     * Returns the description of the module.
+     */
+    public function getDescription( $sDescription ) {
+        return __( 'Triggers actions only once, used mostly by threads.', 'task-scheduler' );
+    }        
+        
+    /**
+     * Deletes stored tasks.
+     */
+    public function doAfterAction( $oRoutine, $sExitCode ) {
 
-		if ( 'NOT_DELETE' === $sExitCode ) { 	
-			return; 
-		}	
-		$vRet = $oRoutine->delete();
-		
-	}	
-	
+        if ( 'NOT_DELETE' === $sExitCode ) {     
+            return; 
+        }    
+        
+        // If the routine status is not changed, it means the routine wants to wait for threads to complete. In that case, the routine should be kept yet.
+        if ( 'processing' === $oRoutine->_routine_status ) {
+            return;
+        }
+        
+        $vRet = $oRoutine->delete();
+        
+    }    
+    
 }
